Niche overlap null model. Resampling through rosario algorithm to preserve the temporal autocorrelation.

Niche overlap null model. Resampling through rosario algorithm to preserve the temporal autocorrelation.

Usage

get_null_model(mat, method, nsim = 100, parallel = FALSE)

Arguments

mat

A matrix with species per time columns.

method

A measure of temporal niche overlap between pairs of species

nsim

The number of simulations

parallel

Logical. Do the resampling on parallel

Value

A matrix with pairwise overlaping temporal resource use.

Examples

temp_overlap(ex1, method = "pianka")
#>    pianka 
#> 0.9975192 

get_null_model(ex1, method = "pianka", nsim = 10)
#> $observed_niche_overlap
#>    pianka 
#> 0.9975192 
#> 
#> $p_value
#> 
#> 	One Sample t-test
#> 
#> data:  res[[1]]
#> t = -19.972, df = 9, p-value = 9.191e-09
#> alternative hypothesis: true mean is not equal to 0.9975192
#> 95 percent confidence interval:
#>  0.05224586 0.24459198
#> sample estimates:
#> mean of x 
#> 0.1484189 
#> 
#> 
#> $null_niche_overlap
#> # A tibble: 10 × 1
#>    pianka
#>     <dbl>
#>  1 0.0284
#>  2 0.0340
#>  3 0.0239
#>  4 0.374 
#>  5 0.296 
#>  6 0.0311
#>  7 0.0528
#>  8 0.299 
#>  9 0.160 
#> 10 0.185 
#>