Random sampling a niche overlap matrix preserving the autocorrelation temporal structure. Each row of the matrix is permuted starting in a random column. The result also could be reversed with a binomial probability of 0.5.

Random sampling a niche overlap matrix preserving the autocorrelation temporal structure. Each row of the matrix is permuted starting in a random column. The result also could be reversed with a binomial probability of 0.5.

Usage

rosario_sample(mat)

Arguments

mat

A matrix with species (rows) and temporal intervals (columns) to be sampled preserving the temporal autocorrelation of the intervals

Value

A random matrix with same number of species and temporal intervals

Examples

rosario_sample(ex1)
#>      INT1 INT2 INT3 INT4 INT5 INT6 INT7 INT8 INT9 INT10 INT11 INT12
#> [1,]    0    0    0    0    0    0    0   13   75    12     0     0
#> [2,]    0    0    0    0    0    7   78   15    0     0     0     0
#> [3,]    0    9   79   12    0    0    0    0    0     0     0     0
#> [4,]    0    0    0   10   80   10    0    0    0     0     0     0
#> [5,]    0    0    7   81   12    0    0    0    0     0     0     0